@@ -28,16 +28,16 @@
         "moment": "^2.30.1",
         "monaco-editor": "^0.45.0",
         "prop-types": "^15.8.1",
-        "react": "^16.14.0",
+        "react": "^18.3.1",
         "react-chartjs-2": "^2.11.2",
         "react-datepicker": "^4.25.0",
-        "react-dom": "^16.14.0",
+        "react-dom": "^18.3.1",
         "react-markdown": "^8.0.7",
         "react-monaco-editor": "^0.50.1",
         "react-router-dom": "^4.2.2",
         "remark-gfm": "^3.0.0",
         "superagent": "^8.1.2",
-        "swagger-ui-react": "^5.10.0"
+        "swagger-ui-react": "^5.17.3"
     },
     "devDependencies": {
         "@babel/core": "^7.22.11",
@@ -53,7 +53,7 @@
         "@types/react": "^16.8.5",
         "@types/react-autocomplete": "^1.8.10",
         "@types/react-datepicker": "^4.19.5",
-        "@types/react-dom": "^16.8.2",
+        "@types/react-dom": "^18.3.0",
         "@types/react-form": "^2.16.1",
         "@types/react-helmet": "^6.1.11",
         "@types/react-router-dom": "^4.2.3",
