@@ -145,8 +145,9 @@ createConnectionSocket(const Network::Address::InstanceConstSharedPtr& peer_addr
   }
   auto connection_socket = std::make_unique<Network::ConnectionSocketImpl>(
       Network::Socket::Type::Datagram, local_addr, peer_addr, Network::SocketCreationOptions{});
+  connection_socket->setDetectedTransportProtocol("quic");
   if (!connection_socket->isOpen()) {
-    ENVOY_LOG_MISC(error, "Failed to create socket");
+    ENVOY_LOG_MISC(error, "Failed to create quic socket");
     return connection_socket;
   }
   connection_socket->addOptions(Network::SocketOptionFactory::buildIpPacketInfoOptions());
