@@ -192,28 +192,34 @@ func (m *manager) Initialize(ctx context.Context) error {
 func (m *manager) Run(ctx context.Context) error {
 	defer m.client.Release()
 
-	err := util.RunTasks(ctx,
-		m.runSynchronizer,
-		m.runSyncSVIDs,
-		m.runSVIDObserver,
-		m.runBundleObserver,
-		m.svid.Run)
+	for {
+		err := util.RunTasks(ctx,
+			m.runSynchronizer,
+			m.runSyncSVIDs,
+			m.runSVIDObserver,
+			m.runBundleObserver,
+			m.svid.Run)
 
-	switch {
-	case err == nil || errors.Is(err, context.Canceled):
-		m.c.Log.Info("Cache manager stopped")
-		return nil
-	case nodeutil.ShouldAgentReattest(err):
-		m.c.Log.WithError(err).Warn("Agent needs to re-attest; removing SVID and shutting down")
-		m.deleteSVID()
-		return err
-	case nodeutil.ShouldAgentShutdown(err):
-		m.c.Log.WithError(err).Warn("Agent is banned: removing SVID and shutting down")
-		m.deleteSVID()
-		return err
-	default:
-		m.c.Log.WithError(err).Error("Cache manager crashed")
-		return err
+		switch {
+		case err == nil || errors.Is(err, context.Canceled):
+			m.c.Log.Info("Cache manager stopped")
+			return nil
+		case nodeutil.ShouldAgentReattest(err):
+			m.c.Log.WithError(err).Warn("Agent needs to re-attest; will attempt to re-attest")
+			reattestError := m.svid.Reattest(ctx)
+			if reattestError != nil {
+				m.c.Log.WithError(reattestError).Error("Agent failed re-attestation; removing SVID and shutting down")
+				m.deleteSVID()
+				return err
+			}
+		case nodeutil.ShouldAgentShutdown(err):
+			m.c.Log.WithError(err).Warn("Agent is banned: removing SVID and shutting down")
+			m.deleteSVID()
+			return err
+		default:
+			m.c.Log.WithError(err).Error("Cache manager crashed")
+			return err
+		}
 	}
 }
 
